Use a uinput device instead of injecting into an existing node
When injecting into an existing node, the events must match the
configuration of that node.  For example, if that node reports itself as
multi-touch, then we'd need to generate multi-touch events.  It's much
safer to create our own input device, configured as we want.

Here we're emulating a touch screen device, reporting only ABS_X/Y and
INPUT_PROP_DIRECT